requirements.txt
discord.py
python-twitter
